rename back to state_machine
